requirements.txt
dj-database-url==0.3.0
Django==1.8.3
djangorestframework==3.1.3
djangorestframework-jwt==1.6.0
gunicorn==19.3.0
newrelic==2.52.0.40
psycopg2==2.6.1
PyJWT==1.4.0
whitenoise==2.0.2
